+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 CHANGELOG
 =========
 
+1.0.1
+-----
+
+ * Change: Allow StatusMessage to be marked as shown
+ * Fix: StatusBusAwareTrait will set StatusBus to null without initialisation
+
+
 1.0.0
 -----